refactor: move NavigationItem to 'navigation'
